+++ vm.c
@@ -33,7 +33,7 @@ alloc:
     return 0;
   // Make sure all those ELO_V bits are zero.
   memset(pgtab, 0, PGSIZE);
-  *pde = (uint)pgtab | asid;
+  *pde = (ulong)pgtab | asid;
   return &pgtab[PTX(va)];
 }
 
@@ -41,14 +41,14 @@ alloc:
 // physical addresses starting at pa. va and size might not
 // be page-aligned.
 static int
-mappages(pde_t *pgdir, char asid, void *va, uint size, uint pa, int perm)
+mappages(pde_t *pgdir, char asid, void *va, ulong size, ulong pa, int perm)
 {
   char *a, *last;
   pte_t *pte;
-  uint elo;
+  ulong elo;
   
-  a = (char*)PGROUNDDOWN((uint)va);
-  last = (char*)PGROUNDDOWN(((uint)va) + size - 1);
+  a = (char*)PGROUNDDOWN((ulong)va);
+  last = (char*)PGROUNDDOWN(((ulong)va) + size - 1);
   for(;;){
     if((pte = walkpgdir(pgdir, asid, a, 1, 1)) == 0)
       return -1;
@@ -89,8 +89,8 @@ mappages(pde_t *pgdir, char asid, void *va, uint size, uint pa, int perm)
 // every process's page table.
 static struct kmap {
   void *virt;
-  uint phys_start;
-  uint phys_end;
+  ulong phys_start;
+  ulong phys_end;
   int perm;
 } kmap[] = {
  { (void*)KERNBASE, 0,             EXTMEM,    ELO_G | ELO_D}, // I/O space
@@ -113,7 +113,7 @@ setupkvm(void)
     panic("PHYSTOP too high");
   for(k = kmap; k < &kmap[NELEM(kmap)]; k++)
     if(mappages(pgdir, 0, k->virt, k->phys_end - k->phys_start, 
-                (uint)k->phys_start, k->perm) < 0)
+                (ulong)k->phys_start, k->perm) < 0)
       return 0;
   return pgdir;
 }
@@ -155,7 +155,7 @@ switchuvm(struct proc *p)
 // Load the initcode into address 0 of pgdir.
 // sz must be less than a page.
 void
-inituvm(pde_t *pgdir, char asid, char *init, uint sz)
+inituvm(pde_t *pgdir, char asid, char *init, ulong sz)
 {
   char *mem;
   
@@ -170,12 +170,12 @@ inituvm(pde_t *pgdir, char asid, char *init, uint sz)
 // Load a program segment into pgdir.  addr must be page-aligned
 // and the pages from addr to addr+sz must already be mapped.
 int
-loaduvm(pde_t *pgdir, char *addr, struct inode *ip, uint offset, uint sz)
+loaduvm(pde_t *pgdir, char *addr, struct inode *ip, ulong offset, ulong sz)
 {
-  uint i, pa, n;
+  ulong i, pa, n;
   pte_t *pte;
 
-  if((uint) addr % PGSIZE != 0)
+  if((ulong) addr % PGSIZE != 0)
     panic("loaduvm: addr must be page aligned");
   for(i = 0; i < sz; i += PGSIZE){
     if((pte = walkpgdir(pgdir, 0, addr+i, 0, 0)) == 0)
@@ -194,10 +194,10 @@ loaduvm(pde_t *pgdir, char *addr, struct inode *ip, uint offset, uint sz)
 // Allocate page tables and physical memory to grow process from oldsz to
 // newsz, which need not be page aligned.  Returns new size or 0 on error.
 int
-allocuvm(pde_t *pgdir, char asid, uint oldsz, uint newsz)
+allocuvm(pde_t *pgdir, char asid, ulong oldsz, ulong newsz)
 {
   char *mem;
-  uint a;
+  ulong a;
 
   if(newsz >= KERNBASE)
     return 0;
@@ -223,10 +223,10 @@ allocuvm(pde_t *pgdir, char asid, uint oldsz, uint newsz)
 // need to be less than oldsz.  oldsz can be larger than the actual
 // process size.  Returns the new process size.
 int
-deallocuvm(pde_t *pgdir, uint oldsz, uint newsz)
+deallocuvm(pde_t *pgdir, ulong oldsz, ulong newsz)
 {
   pte_t *pte;
-  uint a, pa;
+  ulong a, pa;
 
   if(newsz >= oldsz)
     return oldsz;
@@ -253,7 +253,7 @@ deallocuvm(pde_t *pgdir, uint oldsz, uint newsz)
 void
 freevm(pde_t *pgdir)
 {
-  uint i;
+  ulong i;
 
   if(pgdir == 0)
     panic("freevm: no pgdir");
@@ -278,11 +278,11 @@ clearpteu(pde_t *pgdir, char *uva)
 // Given a parent process's page table, create a copy
 // of it for a child.
 pde_t*
-copyuvm(pde_t *pgdir, char asid, uint sz)
+copyuvm(pde_t *pgdir, char asid, ulong sz)
 {
   pde_t *d;
   pte_t *pte;
-  uint pa, i, flags;
+  ulong pa, i, flags;
   char *mem;
 
   if((d = setupkvm()) == 0)
@@ -326,14 +326,14 @@ uva2ka(pde_t *pgdir, char *uva)
 // Most useful when pgdir is not the current page table.
 // uva2ka ensures this only works for PTE_U pages.
 int
-copyout(pde_t *pgdir, uint va, void *p, uint len)
+copyout(pde_t *pgdir, ulong va, void *p, ulong len)
 {
   char *buf, *pa0;
-  uint n, va0;
+  ulong n, va0;
 
   buf = (char*)p;
   while(len > 0){
-    va0 = (uint)PGROUNDDOWN(va);
+    va0 = (ulong)PGROUNDDOWN(va);
     pa0 = uva2ka(pgdir, (char*)va0);
     if(pa0 == 0)
       return -1;
